requirements.txt
# Core dependencies for Streamlit app
streamlit>=1.27.0,<1.46.0
requests>=2.31.0
nest-asyncio>=1.5.8

# For Hugging Face API integration
huggingface-hub>=0.19.0

# For vector embeddings
sentence-transformers

# For data handling
pandas>=2.0.0
numpy>=1.24.0

# For environment variables
python-dotenv>=1.0.0

# For LangChain RAG + Milvus/Zilliz support
langchain>=0.1.0
langchain-community>=0.2.10
langchain-huggingface
langchain-milvus>=0.1.0
pymilvus>=2.3.4 